A fairly significant refactor to support the idea of polyfill 'variants' (see #59).
The interleaving of resolving dependencies and choosing the correct variant for the UA was a major headache and cost me most of the afternoon, and I'm not totally happy with the result.  The difficulty is that each variant may have different dependencies, so you can't resolve dependencies until you've selected a variant.  The two therefore have to alternate as we descend the dependency tree.