Updated localStorage detect.js for Safari
Hi,

On Safari private browsing Polyfill detect the `localStorage` availability with `'localStorage' in this` because `localStorage` exist and is an object. But Safari private browsing limit the localStorage session to 0Mb and when try to set a variable will show an error about the limit size.

With this changes Polyfill detects the Safari private browsing localStorage support.

Many thanks.