Map/Set IteratorResults now always have a value (#901)
IteratorResults returned from `map.keys()`, `map.entries()`, and
`map.values()` used to return an object `{done: true}`, with no `value`
property, when they were done. However, that triggered a bug in
`Array.from` implementations that didn't conform to the spec. This bug
exists in native and polyfilled implementations--including [this one][1]!

The bug causes Array.from to convert an affected iterator into an empty
array, [demonstrated here][2].

It IS a bug in Array.from and not in this implementation, in fact: the
[IteratorResult][3] interface specifies that the `value` property
"may be absent from the conforming object", but unfortunately, both
native and polyfilled implementations of Array.from don't respect that
(at least in Chrome, which is a popular browser, I'm led to believe.)

[1]: https://github.com/Financial-Times/polyfill-service/blob/master/polyfills/Array/from/polyfill.js#L16
[2]: http://jsbin.com/jaliguyaqe/3/edit?html,js,console
[3]: http://www.ecma-international.org/ecma-262/6.0/#sec-iteratorresult-interface