Add compute-at-edge backend for 1/1000 v3/polyfill requests
The code for the compute-at-edge implementation is at https://github.com/Financial-Times/polyfill-c-e

We have ported the VCL to JavaScript using Fastly'sCompute@Edge
platform. The port currently only contains the logic for the /v3/polyfill
endpoint.

We are going to roll out this change to 1 in 1000 requests by default.

Websites can opt-out from using the compute-at-edge implementation
by setting the querystring parameter `use-compute-at-edge-backend` to `no`.

Websites can opt-in to always using the compute-at-edge implementation
by setting the querystring parameter `use-compute-at-edge-backend` to `yes`.

When the compute-at-edge implementation is used, we will not cache the
response within the VCL cache because the compute-at-edge implementation
will handle it's own cache.

If the compute-at-edge implementation returns an error or is not
responding, the VCL implementation will automatically be used instead.